fix(scan): extract ScanImport types module (CQ-003 step 1/4)
Closes #119 step 1. v2 teardown CQ-003.

ScanImport.tsx is 731 lines. Plan splits into co-located sub-
components (ScanImportSession, ScanImportRowEditor, ScanImportSubmit)
under a folder + index.tsx form. This PR is the first step: lift the
shared types and tiny utilities into ./ScanImport/types.ts so the
upcoming sub-components can import them without re-declaring.

Lifted:
  - Row, LookupState, ImportResultRow, ImportResponse types
  - PROVIDER_LABEL constant
  - SCAN_IMPORT_SYMBOLOGIES (the load-bearing module-level symbology
    constant — inline arrays would tear down the multi-MB scanner SDK
    on every render; FE CRIT-2)
  - manufacturerMatches and newRowId helpers

ScanImport.tsx imports them from ./ScanImport/types instead of
declaring them locally. Zero behaviour change. Same scan-import flow,
same dedup keys, same API call shape.

Remaining steps (deferred to follow-up PRs):
  - extract camera + lookupMpnWithRetry into ScanImportSession.tsx
  - extract ScanCard / BagRescanCard / FoundDetails into
    ScanImportRowEditor.tsx
  - extract storage picker + CTA + summary into ScanImportSubmit.tsx
  - finally rename ScanImport.tsx -> ScanImport/index.tsx
